Move series searh text methods into a module
