+++ app/modules/series_search_text.rb
@@ -1,0 +1,23 @@
+module SeriesSearchText
+
+  def title
+    "MARVEL SERIES"
+  end
+
+  def search_title
+    "SEARCH BY TITLE"
+  end
+
+  def search_path
+    "/series"
+  end
+
+  def search_results_title
+    "SERIES SEARCH RESULTS"
+  end
+
+  def search_placeholder
+    "SERIES TITLE"
+  end
+
+end
